Make internal system include directories available.
When running in a Yocto SDK environment pvs-studio could not find internal
system headers while generating dependencies using gcc. Exact origin of
the issue is unknown. Probably has to do with the special environment
the sourced Yocto setup script generates.

Prefixing the system includes in the CXX_FLAGS and C_FLAGS as in this
patch makes the problem disappear.

Signed-off-by: Marc Bodmer <marc.bodmer@securiton.ch>